Move controller and integration test into 'cas2' packages
We need to qualify the `@Service` annotation in the
controller (renamed from `Cas2ApplicationsController` to
`cas2/ApplicationsController`) with the full class path, to
avoid a `ConflictingBeanDefinitionException`.

The alternative to this is to leave the class name as
`Cas2ApplicationsController`, i.e.

`cas2/Cas2ApplicationsControllers`

I'm not sure which approach is going to be clearer and more
reliable.
